requirements.txt
python-telegram-bot[job-queue]==21.7
APScheduler==3.10.4
Pillow>=10.0.0
